Fix: api notes for global topic
